ramips: mt7621: add support for JDCloud RE-SP-01B
JDCloud RE-SP-01B is a dual-band WiFi 5 router based on the MT7621AT.

Specifications:
- SoC: MediaTek MT7621AT
- RAM: 512MB DDR3
- Flash: 32MB SPI NOR
- WiFi: MediaTek MT7603EN (2.4GHz), MediaTek MT7615N (5GHz)
- Ethernet: 1x WAN, 2x LAN (Gigabit Ethernet)
- LEDs: red, blue, green (GPIO controlled)
- Button: Reset (GPIO controlled)
- eMMC: Single onboard (32GB/64GB/128GB)
- USB: 1x USB 2.0 port

MAC Address Structure:
The MAC address shares the structure DC:D8:7C:XX:XX:XX, where:
- WAN, LAN, and 2.4GHz WiFi interfaces use the same MAC as the label.
- 5GHz WiFi uses an offset of the label MAC + 0x800000.

The manufacturer writes the MAC address at different offsets
depending on the storage version of the device:

128GB version:  config + 0x442a
64GB version:   config + 0x4429

As a result, the MAC address of the device cannot be resolved
using DTS alone, and `get_mac_ascii()` is used here to search
for the correct MAC address.
Ref:
https://github.com/openwrt/openwrt/pull/17409#discussion_r1899674262
https://github.com/immortalwrt/immortalwrt/commit/c0c480d

Onboard eMMC:
The eMMC requires the downstream `kmod-sdhci-mt7620` driver due to
upstream `kmod-mmc-mtk` incompatibility.

Flash Instruction:
A third-party bootloader is required to boot the image. You may
need a "Clip" tool to flash the image/bootloader through SPI.

Note that, the official bootloader does provide a web interface
for recovering an official image though. To access it:
  1. Hold the reset button and power on the device.
  2. Set the IP address to 192.168.68.2 on your machine, then
     visit http://192.168.68.1 in a web browser.

Co-authored-by: Chukun Pan <amadeus@jmu.edu.cn>
Signed-off-by: Jin Yijie <jinyijie@outlook.com>